GH-1723: Allow for overriding the default dispatch choice
