+++ jena-fuseki2/jena-fuseki-core/src/main/java/org/apache/jena/fuseki/server/DataService.java
@@ -64,6 +64,8 @@ public class DataService {
     private final AtomicBoolean offlineInProgress       = new AtomicBoolean(false);
     private final AtomicBoolean acceptingRequests       = new AtomicBoolean(true);
 
+    private DispatchFunction plainOperationChooser;
+
     /** Builder for a new DataService. */
     public static Builder newBuilder() { return new Builder(); }
 
@@ -74,14 +76,18 @@ public class DataService {
 
     /** Return a new builder, populated by an existing DatasetService */
     public static Builder newBuilder(DataService dSrv) {
-        return new Builder(dSrv.dataset, dSrv.endpoints, dSrv.operationsMap, dSrv.authPolicy);
+        return new Builder(dSrv.dataset, dSrv.endpoints, dSrv.operationsMap, dSrv.plainOperationChooser, dSrv.authPolicy);
     }
 
     /** Create a {@code DataService} for the given dataset. */
-    private DataService(DatasetGraph dataset, Map<String, EndpointSet> endpoints, ListMultimap<Operation, Endpoint> operationsMap, AuthPolicy authPolicy) {
+    private DataService(DatasetGraph dataset, Map<String, EndpointSet> endpoints,
+                        ListMultimap<Operation, Endpoint> operationsMap,
+                        DispatchFunction plainOperationChooser,
+                        AuthPolicy authPolicy) {
         this.dataset = dataset;
         this.endpoints = Map.copyOf(endpoints);
         this.operationsMap = ArrayListMultimap.create(operationsMap);
+        this.plainOperationChooser = plainOperationChooser;
         this.authPolicy = authPolicy;
         counters.add(CounterName.Requests);
         counters.add(CounterName.RequestsGood);
@@ -153,14 +159,6 @@ public class DataService {
         return operationsMap.keySet().contains(operation);
     }
 
-    public boolean allowUpdate()    { return true; }
-
-    public void goOffline() {
-        offlineInProgress.set(true);
-        acceptingRequests.set(false);
-        state = OFFLINE;
-    }
-
     /** Set any {@link ActionService} processors that are currently unset. */
     public void setEndpointProcessors(OperationRegistry operationRegistry) {
         // Make sure the processor is set for each endpoint.
@@ -179,11 +177,32 @@ public class DataService {
         });
     }
 
+    /**
+     * Decision function for this DataService.
+     * <p>
+     * This is the handler for the case of "no query string, no registered
+     * content-type" that occurs when a name has multiple choices in the dispatcher
+     * so the operation can not be distinguished.
+    * <p>
+     * If null, the system default (in {@code Dispatcher.selectPlainOperation}) is
+     * a quads performed on the dataset - {@link Operation#GSP_R} or
+     * {@link Operation#GSP_RW}.
+     */
+    public DispatchFunction getDefaultOperationChooser() {
+        return plainOperationChooser;
+    }
+
     public void goActive() {
         ensureEnpointProcessors();
         offlineInProgress.set(false);
         acceptingRequests.set(true);
         state = ACTIVE;
+    }
+
+    public void goOffline() {
+        offlineInProgress.set(true);
+        acceptingRequests.set(false);
+        state = OFFLINE;
     }
 
     public boolean isAcceptingRequests() {
@@ -284,17 +303,22 @@ public class DataService {
 
         private Map<String, EndpointSet> endpoints              = new HashMap<>();
         private ListMultimap<Operation, Endpoint> operationsMap = ArrayListMultimap.create();
+        private DispatchFunction plainOperationChooser  = null;
 
         // Dataset-level authorization policy.
         private AuthPolicy authPolicy = null;
 
         private Builder() {}
 
-        private Builder(DatasetGraph dataset, Map<String, EndpointSet> endpoints, ListMultimap<Operation, Endpoint> operationsMap,AuthPolicy authPolicy) {
+        private Builder(DatasetGraph dataset, Map<String, EndpointSet> endpoints,
+                        ListMultimap<Operation, Endpoint> operationsMap,
+                        DispatchFunction plainOperationChooser,
+                        AuthPolicy authPolicy) {
             this();
             this.dataset = dataset;
             this.endpoints.putAll(endpoints);
             this.operationsMap.putAll(operationsMap);
+            this.plainOperationChooser = plainOperationChooser;
             this.authPolicy = authPolicy;
         }
 
@@ -349,13 +373,18 @@ public class DataService {
             operationsMap.remove(endpoint.getOperation(), endpoint);
         }
 
+        public Builder setPlainOperationChooser(DispatchFunction plainOperationChooser) {
+            this.plainOperationChooser = plainOperationChooser;
+            return this;
+        }
+
         public Builder setAuthPolicy(AuthPolicy authPolicy) {
             this.authPolicy = authPolicy;
             return this;
         }
 
         public DataService build() {
-            return new DataService(dataset, endpoints, operationsMap, authPolicy);
+            return new DataService(dataset, endpoints, operationsMap, plainOperationChooser, authPolicy);
         }
     }
 }
