+++ jena-fuseki2/jena-fuseki-core/src/main/java/org/apache/jena/fuseki/server/Endpoint.java
@@ -96,7 +96,7 @@ public class Endpoint implements Counters {
     }
 
     public boolean isUnnamed() {
-        return endpointName == null || endpointName.string ==null || endpointName.string.isEmpty();
+        return endpointName == null || endpointName.string == null || endpointName.string.isEmpty();
     }
 
     public String getName() {
