+++ jena-fuseki2/jena-fuseki-core/src/main/java/org/apache/jena/fuseki/server/EndpointSet.java
@@ -25,18 +25,15 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.BiConsumer;
 
 import org.apache.jena.atlas.logging.Log;
-import org.apache.jena.fuseki.servlets.HttpAction;
 
-/** Collection of endpoints for a dispatch point.
- * A dispatch point is a URL name, without queyr string or content-type.
- * {@linkplain Dispatcher#chooseOperation(HttpAction)} looks at request and decides which
- * {@linkplain Operation} is being requested.
- * 
- * {@linkplain Dispatcher#chooseEndpoint} looks at request,
- * and a {@linkplain DataService}
- * and decides the endpoint.
- * 
- * There can be only be one endpoint for each operation in a {@code EndpointSet}.   
+/**
+ * Collection of endpoints for a dispatch point. A dispatch point is a URL name,
+ * without consideration of the query string or content-type.
+ * <p>
+ * An {@code EndpointSet} is a mapping from Operation to endpoint.
+ * There can be only be one endpoint for each operation in a {@code EndpointSet}.
+ * <p>
+ * See {@link Dispatcher} for the dispatch process.
  */
 public class EndpointSet {
     private final String name;
@@ -49,28 +46,28 @@ public class EndpointSet {
         this.name = name;
         this.single = null;
     }
-    
+
     public void put(Endpoint endpoint) {
         if ( name != null ) {
             if ( ! endpoint.getName().equals(name) )
-                Log.warn(EndpointSet.class, "Different endpoint name: set = '"+name+"' : endpoint = '"+endpoint.getName()+"'");    
+                Log.warn(EndpointSet.class, "Different endpoint name: set = '"+name+"' : endpoint = '"+endpoint.getName()+"'");
         } else {
             if ( ! endpoint.isUnnamed() )
                 Log.warn(EndpointSet.class, "Different endpoint name: set = '' : endpoint = '"+endpoint.getName()+"'");
         }
         Operation operation = endpoint.getOperation();
         if ( endpoints.containsKey(operation) ) {
-            Log.warn(EndpointSet.class, "Redefining endpoint for "+operation); 
+            Log.warn(EndpointSet.class, "Redefining endpoint for "+operation);
         }
         Endpoint endpointPrev = endpoints.put(operation, endpoint);
         resetSingle();
     }
-    
+
     public void remove(Endpoint endpoint) {
         endpoints.remove(endpoint.getOperation());
         resetSingle();
     }
-    
+
     public Endpoint get(Operation operation) {
         return endpoints.get(operation);
     }
@@ -78,7 +75,7 @@ public class EndpointSet {
     public boolean contains(Operation operation) {
         return endpoints.containsKey(operation);
     }
-    
+
     public boolean isEmpty() {
         return endpoints.isEmpty();
     }
@@ -88,31 +85,28 @@ public class EndpointSet {
     }
 
     public void forEach(BiConsumer<Operation, Endpoint> action) { endpoints.forEach(action); }
-    
+
     private void resetSingle() {
         if ( endpoints.size() == 1 )
             single = endpoints.values().iterator().next();
-        else {
-            // Set to null IFF not null. Debug point. 
-            if ( single != null)
-                single = null;
-        }
+        else
+            single = null;
     }
 
     public Collection<Endpoint> endpoints() { return endpoints.values(); }
 
     public Collection<Operation> operations() { return endpoints.keySet(); }
-    
+
     /** Get the Endpoint for a singleton EndpointSet */
     public Endpoint getExactlyOne() {
         return single;
     }
-    
+
     @Override
     public String toString() {
         String x = (name==null)?"":name;
         StringJoiner sj = new StringJoiner(", ", "[", "]");
         operations().forEach(op->sj.add(op.toString()));
-        return x+sj.toString(); 
+        return x+sj.toString();
     }
 }
